180412 노트,선택씬
1. 곡선택 화면 및 곡 선택 (완료)
2. 노트 밀림 현상 해결
- 파싱할 때 float값을 1000을 곱해 소수 3자리까지만 보존하고 있음
- 노트 생성시 최종적으로 넣어줄 때만 1000을 곱해서 소수자리를 최대한 보존.

+++ RhythmGameProjectVer4.2/RhythmGameProject/TrackManager.h
@@ -25,7 +25,7 @@ typedef struct sNoteLine
 typedef struct sNoteInfo
 {
 	char note[3];
-	int startTick;
+	float startTick;
 	int durationTick;
 	int isLongNote;
 	int barNum;
